V3.1 with new control methods
From previously versions, we used long-press to control the lights. It was found that it was hard to hold three or four sides at the same time.

Therefore, sensors control has changed from long-press to short-touch. Each sensor has two status: 1 for on and 0 for off.

Notes: In order to obtain the rainbow effect, if the white side is on, it needs to be turned off.